refactor: middlewware with ts
