--- app/middleware/errorhandler.js
@@ -1,21 +1,0 @@
-'use strict';
-
-module.exports = (option, app) => {
-  return async function(ctx, next) {
-    try {
-      await next();
-    } catch (err) {
-      app.emit('error', err, this);
-      const status = err.status || 500;
-      // 生产环境时 500 错误的详细错误内容不返回给客户端
-      const error = status === 500 && app.config.env === 'prod'
-        ? 'Internal Server Error'
-        : err.message;
-      ctx.body = { error };
-      if (status === 422) {
-        ctx.body.detail = err.errors;
-      }
-      ctx.status = status;
-    }
-  };
-};
