+++ tslint.json
@@ -1,24 +1,50 @@
 {
-    "extends": "tslint:latest",
-    "rules": {
-      "quotemark": [true, "single", "jsx-double"],
-      "no-console": [true, "dir", "log", "error", "warn"],
-      "space-before-function-paren": false,
-      "interface-name": [true, "no-prefix"],
-      "arrow-parens": [true, "ban-single-arg-parens"],
-      "adjacent-overload-signatures": true,
-      "member-access": [false],
-      "object-literal-key-quotes": [true, "as-needed"],
-      "member-ordering": [
-        true,
-        {
-          "order": "fields-first"
-        }
-      ],
-      "object-literal-sort-keys": false,
-      "max-classes-per-file": [true, 10],
-      "max-line-length": false,
-      "variable-name": false,
-      "align": [true, "statements"]
-    }
+  "extends": "tslint:latest",
+  "rules": {
+    "quotemark": [
+      true,
+      "single",
+      "jsx-double"
+    ],
+    "no-console": [
+      true,
+      "dir",
+      "log",
+      "error",
+      "warn"
+    ],
+    "space-before-function-paren": false,
+    "interface-name": [
+      false
+    ],
+    "adjacent-overload-signatures": true,
+    "member-access": [
+      false
+    ],
+    "member-ordering": [
+      true,
+      {
+        "order": "fields-first"
+      }
+    ],
+    "object-literal-sort-keys": false,
+    "max-classes-per-file": [
+      true,
+      10
+    ],
+    "variable-name": [
+      true,
+      "allow-leading-underscore"
+    ],
+    "align": [
+      true,
+      "statements"
+    ],
+    "no-shadowed-variable": false,
+    "only-arrow-functions": false,
+    "no-this-assignment": false,
+    "no-submodule-imports": false,
+    "no-string-literal": false,
+    "no-implicit-dependencies": false
   }
+}
